# Trivial optimize on ORC/Parquet string reading