@taiyang-li 2024-08-09
the failed ut seems not related to this pr : https://s3.amazonaws.com/clickhouse-test-reports/67794/f147e5c39e19d1097361571ebedf4507c744c700/integration_tests__aarch64__[2_6].html